+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/DockerContextMojo.java
@@ -49,13 +49,14 @@ public class DockerContextMojo extends JibPluginConfiguration {
     Preconditions.checkNotNull(project);
     Preconditions.checkNotNull(targetDir);
     Preconditions.checkNotNull(from);
+    Preconditions.checkNotNull(from.image);
 
     ProjectProperties projectProperties = new ProjectProperties(project, getLog());
     String inferredMainClass = projectProperties.getMainClass(mainClass);
 
     try {
       new DockerContextGenerator(projectProperties.getSourceFilesConfiguration())
-          .setBaseImage(from)
+          .setBaseImage(from.image)
           .setJvmFlags(jvmFlags)
           .setMainClass(inferredMainClass)
           .generate(Paths.get(targetDir));